fix: defaulting to transcoded stream
If the user is running their stash server instance on a slower device
(like a raspberry pi) then defaulting to a transcoded stream will be
unwatchable and cause future responses to lag (like if they wanted to
switch to a different video).

This commit changes the order of the streams so that the "direct" source
is presented first in the response json.